pyproject.toml
[project]
name = "pbpm-transformer-benchmark"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "pandas-stubs==2.3.0.250703",
    "scipy-stubs==1.16.1.1",
    "tensorflow>=2.20.0",
]
